feat: addLocationModal 모듈화
