+++ components/viewTrip/tripPlan/location/modal/styles.tsx
@@ -105,26 +105,6 @@ const styles = StyleSheet.create({
     fontWeight: "500",
     color: "white",
   },
-  searchResultsContainer: {
-    position: "absolute",
-    top: "100%",
-    left: 0,
-    right: 0,
-    backgroundColor: "white",
-    borderRadius: 8,
-    borderWidth: 1,
-    borderColor: "#DDDDDD",
-    maxHeight: 200,
-    zIndex: 1000,
-    shadowColor: "#000",
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.25,
-    shadowRadius: 3.84,
-    elevation: 5,
-  },
   searchResultsList: {
     padding: 8,
   },
@@ -155,114 +135,6 @@ const styles = StyleSheet.create({
     marginTop: 8,
     color: "#666",
     fontSize: 14,
-  },
-  locationTabs: {
-    flexDirection: "row",
-    marginBottom: 15,
-    flexWrap: "wrap",
-  },
-  locationTab: {
-    paddingHorizontal: 12,
-    paddingVertical: 8,
-    marginRight: 8,
-    marginBottom: 8,
-    backgroundColor: "#f0f0f0",
-    borderRadius: 20,
-  },
-  activeLocationTab: {
-    backgroundColor: "#4A90E2",
-  },
-  locationTabText: {
-    color: "#333",
-  },
-  activeLocationTabText: {
-    color: "#fff",
-  },
-  addLocationButton: {
-    padding: 8,
-  },
-  removeLocationButton: {
-    marginTop: 15,
-    padding: 12,
-    backgroundColor: "#ff4444",
-    borderRadius: 8,
-    alignItems: "center",
-  },
-  removeLocationText: {
-    color: "#fff",
-    fontWeight: "bold",
-  },
-
-  mapSelectButton: {
-    backgroundColor: "#4A90E2",
-    padding: 15,
-    borderRadius: 10,
-    alignItems: "center",
-    marginBottom: 30,
-  },
-  mapSelectButtonText: {
-    color: "white",
-    fontWeight: "bold",
-  },
-  closeMapButton: {
-    position: "absolute",
-    bottom: 20,
-    left: 20,
-    right: 20,
-    backgroundColor: "white",
-    padding: 15,
-    borderRadius: 10,
-    alignItems: "center",
-  },
-  mapSearchContainer: {
-    position: "absolute",
-    top: 70,
-    left: 10,
-    right: 10,
-    zIndex: 10,
-    backgroundColor: "white",
-    borderRadius: 10,
-  },
-  mapSearchInput: {
-    height: 50,
-    borderWidth: 1,
-    borderColor: "#e0e0e0",
-    borderRadius: 10,
-    paddingHorizontal: 15,
-  },
-  mapSearchResultsContainer: {
-    position: "absolute",
-    top: "100%",
-    left: 0,
-    right: 0,
-    maxHeight: 200,
-    backgroundColor: "white",
-    borderWidth: 1,
-    borderColor: "#E0E0E0",
-    borderRadius: 8,
-    shadowColor: "#000",
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.1,
-    shadowRadius: 4,
-    elevation: 3,
-  },
-  mapSearchResultItem: {
-    padding: 15,
-    borderBottomWidth: 1,
-    borderBottomColor: "#F0F0F0",
-  },
-  mapSearchResultText: {
-    fontSize: 16,
-    color: "#333",
-  },
-  mapSearchResultSubtext: {
-    fontSize: 14,
-    color: "#666",
-    marginTop: 5,
-  },
-  addressSubtext: {
-    color: "gray",
-    fontSize: 12,
   },
 
   markerContainer: {
